+++ src/fillers/typedefs.ts
@@ -1,0 +1,4 @@
+import type { Gradient } from '../gradient/Gradient';
+import type { Pattern } from '../Pattern';
+
+export type TFiller = Gradient<'linear'> | Gradient<'radial'> | Pattern;
